+++ stacker_blueprints/rds/base.py
@@ -29,13 +29,12 @@ def validate_storage_type(value):
     return value
 
 
-def validate_db_instance_identifier(value):
-    l = len(value)
-    if l == 0:
+def validate_db_instance_identifier(value, allow_empty=True):
+    if not value and allow_empty:
         # Empty value will pick up default from stackname
         return value
     pattern = r"^[a-zA-Z][a-zA-Z0-9-]*$"
-    if not (0 < l < 64):
+    if not (0 < len(value) < 64):
         raise ValueError("Must be between 1 and 63 characters in length.")
     if not re.match(pattern, value):
         raise ValueError("Must match pattern: %s" % pattern)
@@ -409,7 +408,13 @@ class MasterInstance(BaseRDS):
             "BackupRetentionPeriod": variables["BackupRetentionPeriod"],
             "DBName": variables["DatabaseName"],
             "DBInstanceClass": variables["InstanceType"],
-            "DBInstanceIdentifier": variables["DBInstanceIdentifier"] or self.context.get_fqn(self.name),
+            "DBInstanceIdentifier": (
+                variables["DBInstanceIdentifier"]
+                or validate_db_instance_identifier(
+                    self.context.get_fqn(self.name),
+                    allow_empty=False
+                )
+            ),
             "DBSnapshotIdentifier": self.get_db_snapshot_identifier(),
             "DBParameterGroupName": Ref("ParameterGroup"),
             "DBSubnetGroupName": Ref(SUBNET_GROUP),
@@ -469,7 +474,13 @@ class ReadReplica(BaseRDS):
             "AllowMajorVersionUpgrade": variables["AllowMajorVersionUpgrade"],
             "AutoMinorVersionUpgrade": variables["AutoMinorVersionUpgrade"],
             "DBInstanceClass": variables["InstanceType"],
-            "DBInstanceIdentifier": variables["DBInstanceIdentifier"] or self.context.get_fqn(self.name),
+            "DBInstanceIdentifier": (
+                variables["DBInstanceIdentifier"]
+                or validate_db_instance_identifier(
+                    self.context.get_fqn(self.name),
+                    allow_empty=False
+                )
+            ),
             "DBParameterGroupName": Ref("ParameterGroup"),
             "Engine": self.engine() or variables["Engine"],
             "EngineVersion": variables["EngineVersion"],
@@ -503,7 +514,10 @@ class ClusterInstance(BaseRDS):
             "AllowMajorVersionUpgrade": variables["AllowMajorVersionUpgrade"],
             "AutoMinorVersionUpgrade": variables["AutoMinorVersionUpgrade"],
             "DBInstanceClass": variables["InstanceType"],
-            "DBInstanceIdentifier": variables["DBInstanceIdentifier"] or self.context.get_fqn(self.name),
+            "DBInstanceIdentifier": (
+                variables["DBInstanceIdentifier"]
+                or self.context.get_fqn(self.name)
+            ),
             "DBSnapshotIdentifier": self.get_db_snapshot_identifier(),
             "DBParameterGroupName": Ref("ParameterGroup"),
             "Engine": self.engine() or variables["Engine"],